pyproject.toml
[tool.poetry]
name = "whapi"
version = "0.1.0"
description = ""
authors = ["Velmax Technologies <veltech.accs.sync@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^3.0.2"
python-dotenv = "^1.0.1"
flask-restful = "^0.3.10"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.8.0"
flask-migrate = "^4.0.7"
flask-marshmallow = "^1.2.1"
marshmallow-sqlalchemy = "^1.0.0"
flask-cors = "^4.0.0"
passlib = "^1.7.4"
flask-jwt-extended = "^4.6.0"
simplejson = "^3.19.2"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
